Move arduino usage to water package #38
